requirements.txt
pandas==1.4.3
asgiref==3.7.2
chardet==5.1.0
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3

